+++ backend/network_develop.go
@@ -1,0 +1,5 @@
+//go:build !fairground && !mainnet
+
+package backend
+
+var DefaultNetworks = []DefaultNetwork{}