+++ frontend/automation/e2e/settings.test.ts
@@ -2,12 +2,12 @@ import type { Page } from '@playwright/test'
 import { expect, test } from '@playwright/test'
 
 import cleanup from '../support/cleanup'
-import { waitForNetworkConnected } from '../support/helpers'
 import initApp from '../support/init-app'
 
-const homeSettingsBtn = 'home-settings'
-const settingsForm = 'settings-form'
-const cancelSettingsBtn = 'cancel-settings'
+const homeSettingsBtn = 'a[href="#/settings"]'
+const settingsForm = 'Settings'
+const appSettingsButton = 'a[href="#/settings/app-settings"]'
+const appSettingsForm = 'App settings'
 
 test.describe('settings', () => {
   let page: Page
@@ -15,39 +15,36 @@ test.describe('settings', () => {
     page = await browser.newPage()
     await initApp(page)
     await page.goto('/')
-    await waitForNetworkConnected(page)
-  })
-
-  test('dialog opens and can be closed', async () => {
-    await page.getByTestId(homeSettingsBtn).click()
-    await expect(page.getByTestId(settingsForm)).toBeVisible()
-    await page.getByTestId(cancelSettingsBtn).click()
-    await expect(page.getByTestId(settingsForm)).toBeHidden()
   })
 
   test('saves and reloads', async () => {
-    await page.getByTestId(homeSettingsBtn).click()
+    await page.locator(homeSettingsBtn).click()
     await expect(page.getByTestId(settingsForm)).toBeVisible()
 
+    // change app settings
+    await page.locator(appSettingsButton).click()
+    await expect(page.getByTestId(appSettingsForm)).toBeVisible()
+
     // assert and change log level
-    await expect(page.getByTestId('log-level').last()).toHaveValue('info')
-    await page.getByTestId('log-level').last().selectOption('debug')
+    await page.locator('#logLevel').selectOption('debug')
 
     // change telemetry
     const radioGroupLocator = page.getByRole('radiogroup')
-    await expect(radioGroupLocator.locator('input[value="yes"]')).toBeChecked()
-    await radioGroupLocator.locator('button[value="no"]').click()
+    await radioGroupLocator.getByTestId('no').click()
 
     // submit
     await page.getByTestId('update-settings').click()
 
-    // page should reload and settings form should now not show
-    await expect(page.getByTestId(settingsForm)).toBeHidden()
+    await page.locator(homeSettingsBtn).click()
+    await page.locator(appSettingsButton).click()
 
-    await page.getByTestId(homeSettingsBtn).click()
-    await expect(page.getByTestId('log-level').last()).toHaveValue('debug')
-    await expect(radioGroupLocator.locator('input[value="no"]')).toBeChecked()
-    await page.getByTestId(cancelSettingsBtn).click()
+    expect(
+      await page.evaluate('document.querySelector("#logLevel").value')
+    ).toEqual('debug')
+    await expect(radioGroupLocator.getByTestId('no')).toHaveAttribute(
+      'aria-checked',
+      'true'
+    )
   })
 
   test.afterAll(async () => {