+++ frontend/automation/support/helpers.ts
@@ -11,14 +11,6 @@ export async function getTextFromClipboard(page: Page): Promise<string> {
   return page.evaluate('navigator.clipboard.readText()')
 }
 
-export function isMainnetConfiguration() {
-  return (
-    process.env.VITE_FEATURE_MODE === 'mainnet' ||
-    process.env.VITE_FEATURE_MODE === '' ||
-    process.env.VITE_FEATURE_MODE === undefined
-  )
-}
-
 export async function unlockWallet(
   page: Page,
   name: string,
@@ -30,23 +22,18 @@ export async function unlockWallet(
 
 // This function usage would be probably possible to remove in most of the places
 // after https://github.com/vegaprotocol/vegawallet-desktop/issues/589 is resolved
-export async function waitForNetworkConnected(page: Page) {
-  const timeout = 5 * 1000
+export async function waitForNetworkConnected(page: Page, network?: string) {
+  const timeout = 10 * 1000
   const serviceStatus = page.getByTestId('service-status')
+  const expectedString = `Wallet Service: ${network || 'test'}`
   try {
-    await expect(serviceStatus).not.toContainText(
-      /^(Not running|Loading|Not reachable)$/,
-      {
-        timeout
-      }
-    )
+    await expect(serviceStatus).toContainText(expectedString, {
+      timeout
+    })
   } catch (e) {
     await page.goto('/')
-    await expect(serviceStatus).not.toContainText(
-      /^(Not running|Loading|Not reachable)$/,
-      {
-        timeout: 2 * timeout
-      }
-    )
+    await expect(serviceStatus).toContainText(expectedString, {
+      timeout: 2 * timeout
+    })
   }
 }